* temp:
  Update
  Update - Rank top 3 avatar
  Add - Loading HUD

# Conflicts:
#	NineZeroProject/Service 2.0/SKPropService.m
#	NineZeroProject/UI 2.0/Index/SKHomepageViewController.m
#	NineZeroProject/UI 2.0/Mascot/SKMascotView.m

+++ NineZeroProject/Model/SKModel.h
@@ -201,8 +201,8 @@
 @end
 
 @interface SKDefaultMascotSkill : NSObject;
-@property (nonatomic, assign)   uint64_t      clue_cooling_time;
-@property (nonatomic, assign)   uint64_t      answer_cooling_time;
+@property (nonatomic, assign)   NSInteger   clue_cooling_time;
+@property (nonatomic, assign)   NSInteger   answer_cooling_time;
 @property (nonatomic, copy)     NSString    *clue_used_gold;
 @property (nonatomic, copy)     NSString    *answer_used_gold;
 @property (nonatomic, copy)     NSString    *clue_used_gemstone;
